feat UpDate code

+++ ex6.js
@@ -2,7 +2,7 @@
 // Escreva um programa que solicita o ano de nascimento de uma pessoa e verifica se ela é maior
 // de 18 anos. Exiba uma mensagem indicando se a pessoa é maior de idade ou não.
 
-let ano = parseFloat (prompt("Digite seu ano de nascimento: "))
+let ano = parseInt (prompt("Digite seu ano de nascimento: "))
 
 let idade = (2024 - ano)
